# docs/swagger.yaml
definitions:
...
  shortpress-server_internal_api.RegisterResponseData:
    properties:
      accessToken:
        type: string
      siteId:
        type: string
    type: object
...
  shortpress-server_internal_api.UserLoginByAuthRequest:
    properties:
      Code:
        type: string
      Credential:
        type: string
      SrcType:
        type: string
      Token:
        type: string
    type: object
  shortpress-server_internal_api.UserLoginData:
    properties:
      accessToken:
        type: string
    type: object
  shortpress-server_internal_api.UserLoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  shortpress-server_internal_api.UserLoginResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/shortpress-server_internal_api.UserLoginData'
      info:
        type: string
    type: object
  shortpress-server_internal_api.UserProfileData:
    properties:
      autoUnlock:
        description: Whether auto-unlock is enabled
        type: boolean
      avatarUrl:
        type: string
      email:
        type: string
      id:
        type: string
      loginType:
        description: '0: email, 1: google, 2: facebook, 3: twitter, 4: tiktok'
        type: integer
      nickname:
        type: string
      premiumExpiresAt:
        type: integer
      premiumType:
        description: '0: free, 1: premium ...'
        type: integer
      status:
        type: integer
    type: object
  shortpress-server_internal_api.UserProfileModifyRequest:
    properties:
      autoUnlock:
        description: Use pointer to allow optional field
        type: boolean
      nickname:
        description: Use pointer to allow optional field
        type: string
    type: object
  shortpress-server_internal_api.UserProfileResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/shortpress-server_internal_api.UserProfileData'
      info:
        type: string
    type: object
  shortpress-server_internal_api.UserRegisterRequest:
    properties:
      email:
        type: string
      nickname:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
...
host: localhost:8000
info:
  contact:
    email: support@swagger.io
    name: shortpress
    url: http://www.swagger.io/support
  description: This is shortpress server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Shortpress Server API
  version: "0.8"
paths:
...
  /api/user/login:
    post:
      consumes:
      - application/json
      description: User login with email and password
      parameters:
      - description: Site ID
        in: header
        name: X-Site-Id
        required: true
        type: string
      - description: Login parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/shortpress-server_internal_api.UserLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Returns access token
          schema:
            $ref: '#/definitions/shortpress-server_internal_api.UserLoginResponse'
      summary: User login
      tags:
      - user
  /api/user/login/oauth2:
    post:
      consumes:
      - application/json
      description: User login with oauth2
      parameters:
      - description: Site ID
        in: header
        name: X-Site-Id
        required: true
        type: string
      - description: Login parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/shortpress-server_internal_api.UserLoginByAuthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Returns access token
          schema:
            $ref: '#/definitions/shortpress-server_internal_api.UserLoginResponse'
      summary: User login with oauth2
      tags:
      - user
  /api/user/profile:
    get:
      consumes:
      - application/json
      description: Get user profile information
      parameters:
      - description: Site ID
        in: header
        name: X-Site-Id
        required: true
        type: string
      - description: Bearer user token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Returns user profile
          schema:
            $ref: '#/definitions/shortpress-server_internal_api.UserProfileResponse'
      security:
      - Bearer: []
      summary: Get user profile
      tags:
      - user
  /api/user/profile-modify:
    post:
      consumes:
      - application/json
      description: Modify user profile information (nickname and autoUnlock)
      parameters:
      - description: Site ID
        in: header
        name: X-Site-Id
        required: true
        type: string
      - description: Bearer user token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Profile modification parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/shortpress-server_internal_api.UserProfileModifyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Returns updated user profile
          schema:
            $ref: '#/definitions/shortpress-server_internal_api.UserProfileResponse'
      security:
      - Bearer: []
      summary: Modify user profile
      tags:
      - user
  /api/user/register:
    post:
      consumes:
      - application/json
      description: Register a new user with email and password
      parameters:
      - description: Site ID
        in: header
        name: X-Site-Id
        required: true
        type: string
      - description: Registration parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/shortpress-server_internal_api.UserRegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/shortpress-server_internal_api.RegisterResponseData'
      summary: Register a new user
      tags:
      - user
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"